Adding lego robot files
I have made a few changes to the original lego robot football file. I'm trying to adjust the timings in order to get the robot to follow the black line (see comments). The catcher now works when pressing space bar and the same motor reverses when clicking the "r" key.